Login/Logout now working using Sessions
I got it working, it's going to need review and some clean up of the
errors being reported before pushing to production.
We're going to need a lot of refactoring as well. But it's a start.

To get the table installed locally just run the SQL query in:
lists/login/sql_statements.txt

- Bumping minor to 0.2.0
- Changed includes to require_once
- Moved the error_reporting above the lists.php require statement which now reports all of the
errors we were not seeing locally.